Refactor: app.pyとbackend構成へ移行
