+++ CHANGELOG.md
@@ -4,6 +4,7 @@ System Version: 1.1.11
 
 ## 1.1.11
 
+- 2026-01-14: リファクタリング: `web_server.py` を廃止し、`app.py` と `backend/` へ移設。起動方法を `uvicorn app:app` に更新。
 - 2026-01-14: 改善: インデックス構築時に対象フォルダが見つからない場合は warn ログを出力。
 
 ## 1.1.10
