+++ app.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+from backend.app_factory import create_app
+
+SYSTEM_VERSION = "1.1.11"
+# File Version: 1.0.0
+
+app = create_app()
