+++ backend/__init__.py
@@ -1,0 +1,6 @@
+"""Backend package."""
+
+# System Version: 1.1.11
+# File Version: 1.0.0
+
+from .app_factory import create_app  # noqa: F401
