+++ backend/routes.py
@@ -19,10 +19,10 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Dict, List, Tuple
 
-from dotenv import load_dotenv, find_dotenv
+from .config import load_env, parse_configured_folders, parse_host_aliases
 
 SYSTEM_VERSION = "1.1.11"
-# File Version: 1.8.3
+# File Version: 1.0.0
 from fastapi import FastAPI, HTTPException, Query
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import FileResponse, JSONResponse, StreamingResponse
@@ -44,7 +44,7 @@ from pdfminer.layout import LAParams, LTTextBoxHorizontal, LTTextBoxVertical, LT
 from pdfminer.pdfpage import PDFPage
 
 # --- 定数 ---
-BASE_DIR = Path(__file__).resolve().parent
+BASE_DIR = Path(__file__).resolve().parent.parent
 INDEX_DIR = BASE_DIR / "indexes"
 INDEX_DIR.mkdir(exist_ok=True)
 
@@ -2458,57 +2458,8 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-STATIC_DIR = Path("static")
+STATIC_DIR = BASE_DIR / "static"
 app.mount("/static", StaticFiles(directory=STATIC_DIR), name="static")
-
-
-def read_env_raw_value(var_name: str) -> str | None:
-    env_path = find_dotenv()
-    if not env_path:
-        return None
-    try:
-        with open(env_path, "r", encoding="utf-8") as file:
-            for line in file:
-                stripped = line.strip()
-                if not stripped or stripped.startswith("#"):
-                    continue
-                if stripped.lower().startswith("export "):
-                    stripped = stripped[7:].lstrip()
-                if "=" not in stripped:
-                    continue
-                key, value = stripped.split("=", 1)
-                if key.strip() != var_name:
-                    continue
-                value = value.strip()
-                if len(value) >= 2 and value[0] == value[-1] and value[0] in "\"'":
-                    value = value[1:-1]
-                return value
-    except OSError:
-        return None
-    return None
-
-
-had_search_folders = "SEARCH_FOLDERS" in os.environ
-raw_search_folders = read_env_raw_value("SEARCH_FOLDERS")
-load_dotenv()
-if not had_search_folders and raw_search_folders is not None:
-    os.environ["SEARCH_FOLDERS"] = raw_search_folders
-
-
-def parse_host_aliases() -> Dict[str, str]:
-    raw = os.getenv("SEARCH_FOLDER_ALIASES", "")
-    normalized = raw.replace("\n", ";").replace("|", ";").replace(",", ";")
-    aliases: Dict[str, str] = {}
-    for part in normalized.split(";"):
-        entry = part.strip()
-        if not entry or entry.startswith("#") or "=" not in entry:
-            continue
-        key, value = entry.split("=", 1)
-        key = key.strip().lower()
-        value = value.strip()
-        if key and value:
-            aliases[key] = value
-    return aliases
 
 
 def display_path_for_path(path: str, aliases: Dict[str, str]) -> str:
@@ -2529,46 +2480,7 @@ def display_path_for_path(path: str, aliases: Dict[str, str]) -> str:
     return f"{prefix}{alias}{rest or ''}"
 
 
-def parse_configured_folders() -> List[Dict[str, str]]:
-    raw = os.getenv("SEARCH_FOLDERS", "")
-    normalized = raw.replace("\n", ";").replace("|", ";").replace(",", ";")
-    folders: List[Dict[str, str]] = []
-    for part in normalized.split(";"):
-        entry = part.strip()
-        if not entry:
-            continue
-        if "#" in entry:
-            entry = entry.split("#", 1)[0].strip()
-        if not entry:
-            continue
-        label = None
-        if "=" in entry:
-            label, entry = entry.split("=", 1)
-            label = label.strip()
-            entry = entry.strip()
-        p = entry.strip().strip('"').strip("'")
-        if not p:
-            continue
-        # smb://host/share または smb:\host\share を UNC/posix 共有パスに変換
-        if p.lower().startswith("smb:"):
-            without_scheme = p[4:].lstrip("/\\")
-            if not without_scheme:
-                continue
-            if os.name == "nt":
-                unc_path = without_scheme.replace("/", "\\")
-                p = f"\\\\{unc_path}"
-            else:
-                posix_path = without_scheme.replace("\\", "/")
-                p = f"//{posix_path}"
-        folders.append(
-            {
-                "path": os.path.abspath(os.path.expanduser(p)),
-                "label": label or "",
-            }
-        )
-    return folders
-
-
+load_env()
 host_aliases = parse_host_aliases()
 configured_folders = parse_configured_folders()
 folder_states: Dict[str, Dict] = {}
@@ -4102,8 +4014,7 @@ def run():
 
     port = int(os.getenv("PORT", "80"))
     cert_dir = os.getenv("CERT_DIR", "certs")
-    base_dir = Path(__file__).resolve().parent
-    cert_path = (base_dir / cert_dir).resolve()
+    cert_path = (BASE_DIR / cert_dir).resolve()
     cert_file = cert_path / "lan-cert.pem"
     key_file = cert_path / "lan-key.pem"
 
@@ -4115,7 +4026,7 @@ def run():
         }
 
     config = uvicorn.Config(
-        "web_server:app",
+        "backend.routes:app",
         host="0.0.0.0",
         port=port,
         reload=False,
